# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  entities.Availability:
    properties:
      chair:
        allOf:
        - $ref: '#/definitions/entities.Chair'
        description: Relacionamentos
      chair_id:
        type: integer
      created_at:
        type: string
      day_of_week:
        description: 0=Domingo, 6=Sábado
        maximum: 6
        minimum: 0
        type: integer
      end_time:
        description: Formato HH:MM
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      start_time:
        description: Formato HH:MM
        type: string
      updated_at:
        type: string
      valid_from:
        type: string
      valid_to:
        type: string
    required:
    - chair_id
    - day_of_week
    - end_time
    - start_time
    type: object
  entities.Booking:
    properties:
      chair:
        $ref: '#/definitions/entities.Chair'
      chair_id:
        type: integer
      created_at:
        type: string
      end_time:
        type: string
      id:
        type: integer
      notes:
        type: string
      start_time:
        type: string
      status:
        enum:
        - agendado
        - confirmado
        - cancelado
        - concluido
        - falta
        type: string
      updated_at:
        type: string
      user:
        allOf:
        - $ref: '#/definitions/entities.User'
        description: Relacionamentos
      user_id:
        type: integer
    required:
    - chair_id
    - end_time
    - start_time
    - user_id
    type: object
  entities.Chair:
    properties:
      availabilities:
        items:
          $ref: '#/definitions/entities.Availability'
        type: array
      bookings:
        description: Relacionamentos
        items:
          $ref: '#/definitions/entities.Booking'
        type: array
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      location:
        maxLength: 100
        minLength: 2
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
      status:
        enum:
        - ativa
        - inativa
        type: string
      updated_at:
        type: string
    required:
    - location
    - name
    type: object
  entities.User:
    properties:
      birth_date:
        type: string
      bookings:
        description: Relacionamentos
        items:
          $ref: '#/definitions/entities.Booking'
        type: array
      cpf:
        type: string
      created_at:
        type: string
      email:
        type: string
      function:
        type: string
      gender:
        enum:
        - masculino
        - feminino
        - outro
        type: string
      id:
        type: integer
      last_login:
        type: string
      name:
        maxLength: 60
        minLength: 2
        type: string
      phone:
        maxLength: 20
        minLength: 10
        type: string
      position:
        type: string
      registration:
        type: string
      requested_role:
        enum:
        - usuario
        - atendente
        - admin
        type: string
      role:
        enum:
        - usuario
        - atendente
        - admin
        type: string
      sector:
        type: string
      status:
        enum:
        - pendente
        - aprovado
        - reprovado
        type: string
      updated_at:
        type: string
    required:
    - cpf
    - email
    - name
    - phone
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API REST para sistema de agendamento de cadeiras de massagem
  title: Sistema de Agendamento API
  version: "1.0"
paths:
...
  /dashboard/admin:
    get:
      consumes:
      - application/json
      description: Retorna dados completos do dashboard com estatísticas, usuários
        pendentes e ocupação das cadeiras para administradores
      parameters:
      - default: hoje
        description: 'Data para filtrar agendamentos (formato: YYYY-MM-DD)'
        in: query
        name: date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Dados completos do dashboard
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Parâmetros inválidos
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Token inválido
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Acesso negado - apenas administradores
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Dashboard do administrador
      tags:
      - dashboard
  /dashboard/attendant:
    get:
      consumes:
      - application/json
      description: Retorna dados do dashboard com agendamentos e estatísticas para
        atendentes
      parameters:
      - default: hoje
        description: 'Data para filtrar agendamentos (formato: YYYY-MM-DD)'
        in: query
        name: date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Dados do dashboard
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Parâmetros inválidos
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Token inválido
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Acesso negado - apenas atendentes e admins
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Dashboard do atendente
      tags:
      - dashboard
  /dashboard/chair-occupancy:
    get:
      consumes:
      - application/json
      description: Retorna dados de ocupação das cadeiras para uma data específica
      parameters:
      - default: hoje
        description: 'Data para calcular ocupação (formato: YYYY-MM-DD)'
        in: query
        name: date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Dados de ocupação das cadeiras
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Formato de data inválido
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Token inválido
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Ocupação das cadeiras
      tags:
      - dashboard
  /dashboard/pending-approvals:
    get:
      consumes:
      - application/json
      description: Retorna lista de usuários aguardando aprovação do administrador
      produces:
      - application/json
      responses:
        "200":
          description: Lista de usuários pendentes
          schema:
            items:
              $ref: '#/definitions/entities.User'
            type: array
        "401":
          description: Token inválido
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Acesso negado - apenas administradores
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Buscar usuários pendentes
      tags:
      - dashboard
  /dashboard/sessions/date/{date}:
    get:
      consumes:
      - application/json
      description: Retorna agendamentos de uma data específica (apenas atendentes
        e admins)
      parameters:
      - description: 'Data para buscar sessões (formato: YYYY-MM-DD)'
        in: path
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Lista de agendamentos da data
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Formato de data inválido
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Token inválido
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Acesso negado
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Sessões por data
      tags:
      - dashboard
  /dashboard/stats/attendance:
    get:
      consumes:
      - application/json
      description: Retorna estatísticas de comparecimento e cancelamento (apenas atendentes
        e admins)
      produces:
      - application/json
      responses:
        "200":
          description: Estatísticas de comparecimento
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Token inválido
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Acesso negado
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Estatísticas de comparecimento
      tags:
      - dashboard
  /dashboard/test-reminders:
    post:
      consumes:
      - application/json
      description: Envia lembretes para todos os agendamentos do próximo dia (para
        testes)
      produces:
      - application/json
      responses:
        "200":
          description: Lembretes enviados com sucesso
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Não autorizado
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Erro interno do servidor
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Enviar lembretes de teste
      tags:
      - Dashboard
...
securityDefinitions:
  Bearer:
    description: Digite 'Bearer' seguido de um espaço e o token JWT obtido no login
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"